Refactor: split operator_commands.rs monolith into domain-specific modules
Break the 2392-line operator_commands.rs into focused modules while
preserving all public function signatures and re-exporting from the
main module so callers (operator_cli.rs, lib.rs) are unchanged.

New modules (each <=400 LOC):
- operator_commands_terminal.rs: terminal probe, file probe, read probe, recipe run
- operator_commands_engineer.rs: engineer loop probe, engineer read probe
- operator_commands_gym.rs: gym list, run, compare, suite
- operator_commands_meeting.rs: meeting/goal/improvement curation probes
- operator_commands_review.rs: review probe, review read probe

Shared helpers (state_root, validated_runtime_segments, print_*, etc.)
remain in operator_commands.rs as pub(crate) and are imported by
sub-modules.
