+++ src/lib.rs
@@ -39,6 +39,11 @@ pub mod ooda_loop;
 pub mod ooda_scheduler;
 pub mod operator_cli;
 pub mod operator_commands;
+mod operator_commands_engineer;
+mod operator_commands_gym;
+mod operator_commands_meeting;
+mod operator_commands_review;
+mod operator_commands_terminal;
 mod persistence;
 pub mod prompt_assets;
 pub mod reflection;
